backend/requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Flask-Migrate
python-dotenv
cryptography
